requirements.txt

python-dotenv       
openai             
elevenlabs         
SpeechRecognition   
sounddevice         
soundfile         
numpy               
python-docx       


streamlit          
requests         